Continuous integration; commented-out continuous deployment
This PR implements Github Actions for the repository.  There are two `steps` in the PR.  One is wholesale commented out, but ready for use.

&nbsp;

# That which exists

The present `step` is a full build pipeline.  It makes a build under node 18 for each of the three major platforms (ubuntu windows mac,) then ubuntu for every current version back through `node 13`, in reverse order.

The reason for this order is that Windows and Mac runners tend to be much slower on GHA than Ubuntu runners, and newer node tends to be faster than older node, so we run the single fastest one first to get soonest feedback, then slowest-next through the entire batch to get soonest completion.

You can easily change whether node 13 is the starting point by adding or removing entries to/from the matrix at line 20.  Please continue to feed them in before 13 in reverse order, if adding.

For each env, the simple `npm install && npm run build` is issued.  I assume that test is part of build, but haven't looked carefully.

&nbsp;

# That which does not

The commented out step is a full deploy to NPM, as well as the generation of a github release, and the pushing of tags to the repo itself.  

***This step is gated to pushes by the owner*** (line 48.)  Other people merging to main ***will not*** trigger a publish, as a safety measure.

In order for this to work, first you must:

1. Go to npmjs.com and log in
2. Create an "automation token" there
3. Bring the "automation token" to Github 
4. Place it in repository settings, under `secrets` / `actions`, as a `repository secret`, with all surrounding whitespace removed
5. Either use the name `PUBLISH_TOKEN_FOR_GH_CI_CD` there, or change line 83 here to whatever you chose instead

This is a reduced version of [my state machine's action setup](https://github.com/StoneCypher/jssm/blob/main/.github/workflows/nodejs.yml), and [I have been using it for some time](https://github.com/StoneCypher/jssm/actions) with [success](https://npmjs.com/package/jssm)